requirements.txt
django==2.2.16
djangorestframework==3.12.4
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
Pillow==8.3.1
sorl-thumbnail==12.7.0